+++ functional-tests/app/client/redux-hooks/routes.js
@@ -3,6 +3,7 @@ import { Index } from './components/Index'
 import { NamedSelectors } from './components/NamedSelectors'
 import { ArrayWithQueryString } from './components/ArrayWithinQueryString'
 import { BasicViewModelTests } from './components/ViewModel'
+import { CustomAggregateIds } from './components/CustomAggregateIds'
 
 export default [
   {
@@ -24,6 +25,10 @@ export default [
       {
         path: '/redux-hooks/view-model/:runId',
         component: BasicViewModelTests,
+      },
+      {
+        path: '/redux-hooks/custom-aggregate-ids/:testId',
+        component: CustomAggregateIds,
       },
     ],
   },